+++ seller/goods.php
@@ -1,5 +1,5 @@
 <?php
-//大商创网络
+/*高度差网络  禁止倒卖 一经发现停止任何服务https://www.dscmall.cn*/
 function cat_list_one_new($cat_id = 0, $cat_level = 0, $sel_cat)
 {
 	if ($cat_id == 0) {
@@ -140,7 +140,7 @@ $commission_setting = admin_priv('commission_setting', '', false);
 $smarty->assign('commission_setting', $commission_setting);
 if ($_REQUEST['act'] == 'list' || $_REQUEST['act'] == 'trash' || $_REQUEST['act'] == 'no_comment') {
 	admin_priv('goods_manage');
-	$sql = 'DELETE FROM' . $GLOBALS['ecs']->table('products_changelog') . 'WHERE admin_id = \'' . $admin_id . '\'';
+	$sql = 'DELETE FROM' . $GLOBALS['ecs']->table('products_changelog') . ' WHERE goods_id = 0 AND admin_id = \'' . $admin_id . '\'';
 	$GLOBALS['db']->query($sql);
 	get_del_goodsimg_null();
 	get_del_goods_gallery();
@@ -246,8 +246,6 @@ if ($_REQUEST['act'] == 'list' || $_REQUEST['act'] == 'trash' || $_REQUEST['act'
 }
 else {
 	if ($_REQUEST['act'] == 'add' || $_REQUEST['act'] == 'edit' || $_REQUEST['act'] == 'copy') {
-		$sql = 'DELETE FROM' . $GLOBALS['ecs']->table('products_changelog') . 'WHERE admin_id = \'' . $admin_id . '\'';
-		$GLOBALS['db']->query($sql);
 		get_del_goodsimg_null();
 		get_del_goods_gallery();
 		get_del_update_goods_null();
@@ -331,6 +329,14 @@ else {
 		$smarty->assign('grade_rank', $grade_rank);
 		$smarty->assign('integral_scale', $_CFG['integral_scale']);
 		$goods_id = isset($_REQUEST['goods_id']) && !empty($_REQUEST['goods_id']) ? intval($_REQUEST['goods_id']) : 0;
+		$changelog_where = '';
+
+		if (empty($goods_id)) {
+			$changelog_where = ' AND admin_id = \'' . $admin_id . '\'';
+		}
+
+		$sql = 'DELETE FROM' . $GLOBALS['ecs']->table('products_changelog') . (' WHERE goods_id = \'' . $goods_id . '\'') . $changelog_where;
+		$GLOBALS['db']->query($sql);
 
 		if ($is_add) {
 			$res = array();
@@ -375,7 +381,7 @@ else {
 				'user_cat'           => 0,
 				'goods_unit'         => '个',
 				'goods_extend'       => array('is_reality' => 0, 'is_return' => 0, 'is_fast' => 0)
-				);
+			);
 
 			if ($code != '') {
 				$goods['goods_number'] = 0;
@@ -390,7 +396,7 @@ else {
 			$goods_article_list = array();
 			$sql = 'DELETE FROM ' . $ecs->table('goods_article') . (' WHERE goods_id = 0 AND admin_id = \'' . $_SESSION['seller_id'] . '\'');
 			$db->query($sql);
-			$sql = 'DELETE FROM ' . $ecs->table('goods_attr') . ' WHERE goods_id = 0';
+			$sql = 'DELETE FROM ' . $ecs->table('goods_attr') . (' WHERE goods_id = 0 AND admin_id = \'' . $admin_id . '\'');
 			$db->query($sql);
 			$img_list = array();
 		}
@@ -452,7 +458,7 @@ else {
 					'rank_integral'      => 0,
 					'user_cat'           => 0,
 					'goods_extend'       => array('is_reality' => 0, 'is_return' => 0, 'is_fast' => 0)
-					);
+				);
 			}
 
 			$goods['goods_video_path'] = !empty($goods['goods_video']) ? get_image_path($goods['goods_id'], $goods['goods_video']) : '';
@@ -565,7 +571,7 @@ else {
 					$db->autoExecute($ecs->table('goods_article'), $row, 'INSERT');
 				}
 
-				$sql = 'DELETE FROM ' . $ecs->table('goods_attr') . ' WHERE goods_id = 0';
+				$sql = 'DELETE FROM ' . $ecs->table('goods_attr') . (' WHERE goods_id = 0 AND admin_id = \'' . $admin_id . '\'');
 				$db->query($sql);
 				$sql = 'SELECT 0 AS goods_id, attr_id, attr_value, attr_price ' . 'FROM ' . $ecs->table('goods_attr') . ' WHERE goods_id = \'' . intval($_REQUEST['goods_id']) . '\' ';
 				$res = $db->query($sql);
@@ -2236,7 +2242,13 @@ else {
 						}
 					}
 
-					$sql = 'DELETE FROM' . $ecs->table('products_changelog') . ('WHERE goods_id = \'' . $goods_id . '\' AND admin_id = \'') . $_SESSION['seller_id'] . '\'';
+					$products_changelog_where = '';
+
+					if (empty($goods_id)) {
+						$products_changelog_where = ' AND admin_id = \'' . $admin_id . '\'';
+					}
+
+					$sql = 'DELETE FROM' . $ecs->table('products_changelog') . ('WHERE goods_id = \'' . $goods_id . '\'') . $products_changelog_where;
 					$db->query($sql);
 					$goods = get_admin_goods_info($goods_id, array('promote_price', 'promote_start_date', 'promote_end_date', 'user_id', 'model_attr'));
 					if ($GLOBALS['_CFG']['add_shop_price'] == 0 && $goods['model_attr'] == 0) {
